Refactor NSQLookupd to be importable
Refactor NSQLookupd to be importable

Fix formatting, make changes from comments on PR WRT exporting of variables

Rename imports package to nsqlookupd, move nsqlookupd/main.go to apps/nsqlookupd/nsqlookupd.go.

Attempt to fix TravisCI build of nsqlookupd

Change some testing variables back to private

Fixes after rebase

Make TcpAddr and HttpAddr unexported again, remove nsqld prefix from app
