+++ nsqlookupd/client_v1.go
@@ -1,4 +1,4 @@
-package main
+package nsqlookupd
 
 import (
 	"net"
