+++ nsqlookupd/context.go
@@ -1,4 +1,4 @@
-package main
+package nsqlookupd
 
 type Context struct {
 	nsqlookupd *NSQLookupd
