+++ nsqlookupd/registration_db.go
@@ -1,4 +1,4 @@
-package main
+package nsqlookupd
 
 import (
 	"fmt"
