+++ nsqlookupd/registration_db_test.go
@@ -1,4 +1,4 @@
-package main
+package nsqlookupd
 
 import (
 	"github.com/bmizerany/assert"
